compose/locations/requirements.txt
psycopg2-binary
sqlalchemy
requests
gtfs-realtime-bindings
geopy
shapely
geoalchemy2
